# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateReviewRequest:
    properties:
      comment:
        maxLength: 5000
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        maxLength: 150
        type: string
      variant_id:
        type: string
    required:
    - rating
    type: object
...
  dto.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RatingSummaryResponse:
    properties:
      average_rating:
        type: number
      product_id:
        type: string
      rating_breakdown:
        additionalProperties:
          type: integer
        type: object
      total_reviews:
        type: integer
    type: object
...
  dto.ReviewListItemResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      rating:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/dto.ReviewUserResponse'
      variant_id:
        type: string
    type: object
  dto.ReviewListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ReviewListItemResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
  dto.ReviewResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      rating:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      variant_id:
        type: string
    type: object
  dto.ReviewUserResponse:
    properties:
      display_name:
        type: string
      id:
        type: string
    type: object
...
  dto.UpdateReviewRequest:
    properties:
      comment:
        maxLength: 5000
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        maxLength: 150
        type: string
    type: object
...
  dto.UserReviewListItemResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/dto.UserReviewProductResponse'
      rating:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      variant_id:
        type: string
    type: object
  dto.UserReviewListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserReviewListItemResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
  dto.UserReviewProductResponse:
    properties:
      id:
        type: string
      name:
        type: string
      thumbnail_url:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: 'REST API for Product & Category Service. Management routes require
    an Auth0 RS256 bearer access token whose exact permission matches the route; denials
    use the {"error":{"code":"...","message":"...","details":{}}} envelope with 401
    UNAUTHENTICATED for a missing, malformed, invalid, expired, or wrong-audience
    credential and 403 FORBIDDEN for a valid token without the route''s exact permission.
    Public exception: GET /categories, GET /categories/{id}, GET /categories/dropdown,
    GET /products, and GET /products/{id} require no token and ignore any supplied
    Authorization header. Customer exception: POST /products/{id}/reviews, PATCH /reviews/{reviewId},
    DELETE /reviews/{reviewId}, and GET /users/me/reviews keep the existing customer
    HS256 token and ownership rules, and an Auth0 permission is not a substitute for
    customer identity.'
  title: Gin Product Service API
  version: "1.0"
paths:
...
  /products/{id}/reviews:
    get:
      description: Get a paginated, filterable, sortable list of reviews for a product
      parameters:
      - description: Product UUID
        in: path
        name: id
        type: string
      - description: Page number (1-indexed)
        in: query
        name: page
        type: integer
      - description: Page size (max 100)
        in: query
        name: limit
        type: integer
      - description: Filter to a specific star rating (1-5)
        in: query
        name: rating
        type: integer
      - description: Accepted but not filtered (no purchase history data)
        in: query
        name: verified_only
        type: boolean
      - description: Only reviews with a non-empty comment
        in: query
        name: has_comment
        type: boolean
      - description: 'Sort: newest, oldest, highest_rating, lowest_rating'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List reviews for a product
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Create a review for a product, optionally tied to a purchased variant.
        Requires authentication.
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Review to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a review for a product
      tags:
      - reviews
  /products/{id}/reviews/summary:
    get:
      description: Get the aggregate rating stats (average, total, breakdown) for
        a product
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RatingSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a product's rating summary
      tags:
      - reviews
...
  /reviews/{reviewId}:
    delete:
      description: Delete a review. Only the author can delete it. Requires authentication.
      parameters:
      - description: Review UUID
        in: path
        name: reviewId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete your own review
      tags:
      - reviews
    get:
      description: Get a single review by id (same shape as the list item)
      parameters:
      - description: Review UUID
        in: path
        name: reviewId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewListItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a single review
      tags:
      - reviews
    patch:
      consumes:
      - application/json
      description: Update a review's rating/title/comment. Only the author can edit
        it. Requires authentication.
      parameters:
      - description: Review UUID
        in: path
        name: reviewId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update your own review
      tags:
      - reviews
  /users/me/reviews:
    get:
      description: Get a paginated list of the authenticated user's reviews. Requires
        authentication.
      parameters:
      - description: Page number (1-indexed)
        in: query
        name: page
        type: integer
      - description: Page size (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserReviewListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List the current user's reviews
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Auth0 RS256 access token as `Bearer <token>`. The token must carry
      the operation's exact permission and be signed by a key currently published
      by the configured Auth0 domain. Token issuer and audience must match the configured
      Auth0 API; an invalid, expired, or unintended token yields 401.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"